package/network-manager: make crypto library optional
The network-manager package currently pulls in either gnutls or libnss,
neither of which are very common and it might be the only reason why
they are present on a system.

However, most of NetworkManager works just fine without any cryptography
support, it only seems to be used in test cases and 802.1X support code.

Remove the dependency but use a library if it is present.

Note that this changes the default behavior. If network-manager was the
only package pulling in gnutls, it won't do this anymore and use the
"null" backend. Add a note about this to the manual.

Signed-off-by: Florian Larysch <fl@n621.de>

Upstream: https://patchwork.ozlabs.org/project/buildroot/patch/20250531143928.25446-2-fl@n621.de/
(This won't get applied to 2025.02 upstream though, so this is a
backport)